Refactored: file structure
