--- create_py_from.bat
@@ -1,4 +1,0 @@
-pyuic5 can_unit_widget.ui -o can_unit_widget.py
-pyuic5 can_usb_bridge_client_widget.ui -o can_usb_bridge_client_widget.py
-pyuic5 can_usb_bridge_client_win.ui -o can_usb_bridge_client_win.py
-pause
