--- pyinstaller_setup.bat
@@ -1,3 +1,0 @@
-pyinstaller can_unit.py --clean --onefile --noconsole --distpath can_usb_app
-
-pause
