--- pyinstaller_setup_test.bat
@@ -1,3 +1,0 @@
-pyinstaller test.py --clean --onefile --distpath test_apprun
-
-pause
