--- test.py
@@ -1,4 +1,0 @@
-import os
-
-print(os.getcwd())
-print(__file__)
